ajustes e novos tests
- movendo /accounts/ para /web/accounts/
- novos tests e fix no tests que esperava 200 OK na home
- novos tests para acessar /dashboard/ (blueprint com login required)
- novos tests para acessar /scheduler/ (blueprint com login required)
- refactor no create_app mais modularizado
