+++ opac_proc/web/accounts/templates/accounts/register.html
