+++ opac_proc/web/webapp.py
@@ -16,12 +16,32 @@ PROJECT_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), '../..'))
 sys.path.append(PROJECT_PATH)
 
 from opac_proc.web import urls
-from opac_proc.accounts import accounts as accounts_bp
 
 db = MongoEngine()
 toolbar = DebugToolbarExtension()
 login_manager = LoginManager()
+login_manager.session_protection = 'strong'
+login_manager.login_view = 'accounts.login'
 mail = Mail()
+
+
+def register_extensions(app):
+    from accounts.handlers import *   # NOQA
+    login_manager.init_app(app)
+    db.init_app(app)
+    app.session_interface = MongoEngineSessionInterface(db)
+    toolbar.init_app(app)
+    mail.init_app(app)
+
+
+def regiter_bluprints(app):
+    from accounts import accounts
+    from accounts.helpers import check_user_logged_in_or_redirect
+    rq_scheduler_dashboard.blueprint.before_request(check_user_logged_in_or_redirect)
+    rq_dashboard.blueprint.before_request(check_user_logged_in_or_redirect)
+    app.register_blueprint(accounts)
+    app.register_blueprint(rq_scheduler_dashboard.blueprint, url_prefix='/scheduler')
+    app.register_blueprint(rq_dashboard.blueprint, url_prefix='/dashboard')
 
 
 def create_app(test_mode=False):
@@ -41,24 +61,10 @@ def create_app(test_mode=False):
         app.config['TESTING'] = True
         app.config['DEBUG'] = False
 
-    # login
-    from opac_proc.accounts.handlers import *  # NOQA vem aqui: check_user_logged_in_or_redirect
-    login_manager.session_protection = 'strong'
-    login_manager.login_view = 'accounts.login'
-    login_manager.init_app(app)
-
-    db.init_app(app)
-    # http://docs.mongoengine.org/projects/flask-mongoengine/en/latest/#session-interface
-    app.session_interface = MongoEngineSessionInterface(db)
-    toolbar.init_app(app)
-    mail.init_app(app)
-
+    # init extensions
+    register_extensions(app)
     # blueprints
-    rq_scheduler_dashboard.blueprint.before_request(check_user_logged_in_or_redirect)
-    rq_dashboard.blueprint.before_request(check_user_logged_in_or_redirect)
-    app.register_blueprint(accounts_bp)
-    app.register_blueprint(rq_scheduler_dashboard.blueprint, url_prefix='/scheduler')
-    app.register_blueprint(rq_dashboard.blueprint, url_prefix='/dashboard')
+    regiter_bluprints(app)
 
     app.wsgi_app = ProxyFix(app.wsgi_app)
     urls.add_url_rules(app)
